Resolve dependency cycle between `api-client` and `actions-util`
