+++ lib/analyze-action-env.test.js
@@ -49,7 +49,7 @@ const util = __importStar(require("./util"));
         sinon
             .stub(actionsUtil, "createStatusReportBase")
             .resolves({});
-        sinon.stub(actionsUtil, "sendStatusReport").resolves(true);
+        sinon.stub(api, "sendStatusReport").resolves(true);
         sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
         const gitHubVersion = {
             type: util.GitHubVariant.DOTCOM,
