+++ lib/analyze-action.js
@@ -34,6 +34,7 @@ const core = __importStar(require("@actions/core"));
 const actionsUtil = __importStar(require("./actions-util"));
 const analyze_1 = require("./analyze");
 const api_client_1 = require("./api-client");
+const api = __importStar(require("./api-client"));
 const autobuild_1 = require("./autobuild");
 const codeql_1 = require("./codeql");
 const config_utils_1 = require("./config-utils");
@@ -66,10 +67,10 @@ async function sendStatusReport(startedAt, config, stats, error, trapCacheUpload
             trap_cache_upload_duration_ms: Math.round(trapCacheUploadTime || 0),
             trap_cache_upload_size_bytes: Math.round(await (0, trap_caching_1.getTotalCacheSize)(config.trapCaches, logger)),
         };
-        await actionsUtil.sendStatusReport(trapCacheUploadStatusReport);
+        await api.sendStatusReport(trapCacheUploadStatusReport);
     }
     else {
-        await actionsUtil.sendStatusReport(statusReport);
+        await api.sendStatusReport(statusReport);
     }
 }
 exports.sendStatusReport = sendStatusReport;
@@ -145,7 +146,7 @@ async function run() {
     util.initializeEnvironment(actionsUtil.getActionVersion());
     const logger = (0, logging_1.getActionsLogger)();
     try {
-        if (!(await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("finish", "starting", startedAt)))) {
+        if (!(await api.sendStatusReport(await actionsUtil.createStatusReportBase("finish", "starting", startedAt)))) {
             return;
         }
         config = await (0, config_utils_1.getConfig)(actionsUtil.getTemporaryDirectory(), logger);
