+++ lib/autobuild-action.js
@@ -41,7 +41,7 @@ async function sendCompletedStatusReport(startedAt, allLanguages, failingLanguag
         autobuild_languages: allLanguages.join(","),
         autobuild_failure: failingLanguage,
     };
-    await (0, actions_util_1.sendStatusReport)(statusReport);
+    await (0, api_client_1.sendStatusReport)(statusReport);
 }
 async function run() {
     const startedAt = new Date();
@@ -49,7 +49,7 @@ async function run() {
     let currentLanguage = undefined;
     let languages = undefined;
     try {
-        if (!(await (0, actions_util_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)("autobuild", "starting", startedAt)))) {
+        if (!(await (0, api_client_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)("autobuild", "starting", startedAt)))) {
             return;
         }
         const gitHubVersion = await (0, api_client_1.getGitHubVersion)();
