+++ lib/init-action-post.js
@@ -51,7 +51,7 @@ async function runWrapper() {
     catch (unwrappedError) {
         const error = (0, util_1.wrapError)(unwrappedError);
         core.setFailed(error.message);
-        await (0, actions_util_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)("init-post", (0, actions_util_1.getActionsStatus)(error), startedAt, error.message, error.stack));
+        await (0, api_client_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)("init-post", (0, actions_util_1.getActionsStatus)(error), startedAt, error.message, error.stack));
         return;
     }
     const statusReportBase = await (0, actions_util_1.createStatusReportBase)("init-post", "success", startedAt);
@@ -59,7 +59,7 @@ async function runWrapper() {
         ...statusReportBase,
         ...uploadFailedSarifResult,
     };
-    await (0, actions_util_1.sendStatusReport)(statusReport);
+    await (0, api_client_1.sendStatusReport)(statusReport);
 }
 void runWrapper();
 //# sourceMappingURL=init-action-post.js.map
