+++ lib/upload-sarif-action.js
@@ -26,6 +26,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const core = __importStar(require("@actions/core"));
 const actionsUtil = __importStar(require("./actions-util"));
 const actions_util_1 = require("./actions-util");
+const api_client_1 = require("./api-client");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
 const upload_lib = __importStar(require("./upload-lib"));
@@ -36,12 +37,12 @@ async function sendSuccessStatusReport(startedAt, uploadStats) {
         ...statusReportBase,
         ...uploadStats,
     };
-    await actionsUtil.sendStatusReport(statusReport);
+    await (0, api_client_1.sendStatusReport)(statusReport);
 }
 async function run() {
     const startedAt = new Date();
     (0, util_1.initializeEnvironment)((0, actions_util_1.getActionVersion)());
-    if (!(await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("upload-sarif", "starting", startedAt)))) {
+    if (!(await (0, api_client_1.sendStatusReport)(await actionsUtil.createStatusReportBase("upload-sarif", "starting", startedAt)))) {
         return;
     }
     try {
@@ -61,7 +62,7 @@ async function run() {
         const message = error.message;
         core.setFailed(message);
         console.log(error);
-        await actionsUtil.sendStatusReport(await actionsUtil.createStatusReportBase("upload-sarif", actionsUtil.getActionsStatus(error), startedAt, message, error.stack));
+        await (0, api_client_1.sendStatusReport)(await actionsUtil.createStatusReportBase("upload-sarif", actionsUtil.getActionsStatus(error), startedAt, message, error.stack));
         return;
     }
 }
