+++ src/analyze-action-env.test.ts
@@ -29,7 +29,7 @@ test("analyze action with RAM & threads from environment variables", async (t) =
     sinon
       .stub(actionsUtil, "createStatusReportBase")
       .resolves({} as actionsUtil.StatusReportBase);
-    sinon.stub(actionsUtil, "sendStatusReport").resolves(true);
+    sinon.stub(api, "sendStatusReport").resolves(true);
     sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
 
     const gitHubVersion: util.GitHubVersion = {
